feat: create card sub components
